+++ src/arize/_generated/api_client/models/categorical_annotation_config.py
@@ -33,7 +33,7 @@ class CategoricalAnnotationConfig(BaseModel):
     name: StrictStr = Field(description="The name of the annotation config")
     created_at: datetime = Field(description="The timestamp for when the annotation config was created")
     space_id: StrictStr = Field(description="The space id the annotation config belongs to")
-    type: StrictStr = Field(description="The type of the annotation config")
+    type: StrictStr
     values: List[CategoricalAnnotationValue] = Field(description="An array of categorical annotation values")
     optimization_direction: Optional[OptimizationDirection] = None
     __properties: ClassVar[List[str]] = ["id", "name", "created_at", "space_id", "type", "values", "optimization_direction"]